+++ logout.php
@@ -1,7 +1,4 @@
 <?php
-
 session_start();
-
 $_SESSION = [];
-
 header("Location: /index.php");
