+++ sign_in.php
@@ -1,14 +1,14 @@
 <?php
 require_once('bootstrap.php');
-require_once('user_function.php');
 require_once('functions/template.php');
 require_once('functions/subsidiary.php');
 require_once('functions/validate.php');
+require_once('functions/datetime.php');
+require_once('functions/get_from_get_or_post.php');
 
 $errors_validate = [];
 
 if (isset($_POST['submit'])) {  //Если есть такое поле в POST, значит форма отправлена
-
   $errors_validate = validate_sign_in($connection);
 
   //Если были ошибки валидации - возвращаем на страницу добавления нового лота с показом ошибок
